uint32_t->uintptr_t

+++ thread-basics/notification-chain/threaded_subscriber.c
@@ -3,12 +3,13 @@
 #include <memory.h>
 #include <pthread.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/types.h> /*for pid_t*/
 #include <unistd.h>    /*for getpid()*/
 
-void create_subscriber_thread(uint32_t client_id);
+void create_subscriber_thread(uintptr_t client_id);
 
 extern rt_table_t *publisher_get_rt_table();
 
@@ -36,7 +37,7 @@ static void *subscriber_thread_fn(void *arg) {
    rt_entry_keys.mask = 32;
    rt_table_register_for_notification(publisher_get_rt_table(), &rt_entry_keys,
                                       sizeof(rt_entry_keys_t), test_cb,
-                                      (uint32_t)arg);
+                                      (uintptr_t)arg);
 
    /* register for Notif 122.1.1.2/32 */
    memset(&rt_entry_keys, 0, sizeof(rt_entry_keys_t));
@@ -44,7 +45,7 @@ static void *subscriber_thread_fn(void *arg) {
    rt_entry_keys.mask = 32;
    rt_table_register_for_notification(publisher_get_rt_table(), &rt_entry_keys,
                                       sizeof(rt_entry_keys_t), test_cb,
-                                      (uint32_t)arg);
+                                      (uintptr_t)arg);
 
    /* register for Notif 122.1.1.3/32 */
    memset(&rt_entry_keys, 0, sizeof(rt_entry_keys_t));
@@ -52,13 +53,13 @@ static void *subscriber_thread_fn(void *arg) {
    rt_entry_keys.mask = 32;
    rt_table_register_for_notification(publisher_get_rt_table(), &rt_entry_keys,
                                       sizeof(rt_entry_keys_t), test_cb,
-                                      (uint32_t)arg);
+                                      (uintptr_t)arg);
 
    pause();
    return NULL;
 }
 
-void create_subscriber_thread(uint32_t client_id) {
+void create_subscriber_thread(uintptr_t client_id) {
 
    pthread_attr_t attr;
    pthread_t subs_thread;
@@ -67,5 +68,5 @@ void create_subscriber_thread(uint32_t client_id) {
    pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_DETACHED);
 
    pthread_create(&subs_thread, &attr, subscriber_thread_fn, (void *)client_id);
-   printf("Subscriber %u created\n", client_id);
+   printf("Subscriber %lu created\n", client_id);
 }
